Split frontend Submissions/Requests by db
Restructure Ember routes so each db (st26/bioproject/biosample) gets
its own Requests and Submissions listing. URLs follow
/:db/requests/... and /:db/submissions/... mirroring the API.

The top nav becomes two dropdowns (Requests, Submissions), each
listing the three dbs. Detail/list templates now show the db in
their headings, and LinkTos pass db as the first model so resetNamespace
routes (request, submission, update) resolve correctly.

Pagination accepts an optional @models array to be db-aware. The
previous @route="requests.index" form (no controller path under db)
emitted ?page=1 because the active controller for db.requests.index
lives at controllers/db/requests/index.ts; controllers and routes are
moved to keep query-param defaults working.
