+++ web/app/router.ts
@@ -10,21 +10,23 @@ Router.map(function () {
   this.route('login');
   this.route('account');
 
-  this.route('requests', function () {
-    this.route('new');
-    this.route('request', { path: ':request_id', resetNamespace: true });
-  });
+  this.route('db', { path: ':db' }, function () {
+    this.route('requests', function () {
+      this.route('new');
+      this.route('request', { path: ':request_id', resetNamespace: true });
+    });
 
-  this.route('updates', function () {
-    this.route('update', { path: ':update_id', resetNamespace: true });
-  });
+    this.route('updates', function () {
+      this.route('update', { path: ':update_id', resetNamespace: true });
+    });
 
-  this.route('submissions', function () {
-    this.route('submission', { path: ':submission_id', resetNamespace: true }, function () {
-      this.route('accessions');
+    this.route('submissions', function () {
+      this.route('submission', { path: ':submission_id', resetNamespace: true }, function () {
+        this.route('accessions');
 
-      this.route('updates', function () {
-        this.route('new');
+        this.route('updates', function () {
+          this.route('new');
+        });
       });
     });
   });
