+++ web/app/routes/db.ts
@@ -4,10 +4,14 @@ import { service } from '@ember/service';
 import type CurrentUserService from 'repository/services/current-user';
 import type Transition from '@ember/routing/transition';
 
-export default class SubmissionsRoute extends Route {
+export default class DbRoute extends Route {
   @service declare currentUser: CurrentUserService;
 
   beforeModel(transition: Transition) {
     this.currentUser.ensureLogin(transition);
+  }
+
+  model({ db }: { db: string }) {
+    return { db };
   }
 }
