+++ web/app/routes/db/requests/index.ts
@@ -4,7 +4,8 @@ import { service } from '@ember/service';
 import type { RequestManager } from '@warp-drive/core';
 import type { paths } from 'schema/openapi';
 
-type SubmissionSummary = paths['/{db}/submissions']['get']['responses']['200']['content']['application/json'];
+type SubmissionRequestSummary =
+  paths['/{db}/submission_requests']['get']['responses']['200']['content']['application/json'];
 
 export default class extends Route {
   @service declare requestManager: RequestManager;
@@ -16,13 +17,16 @@ export default class extends Route {
   };
 
   async model({ page }: { page?: number }) {
-    const { content, response } = await this.requestManager.request<SubmissionSummary>({
-      url: '/st26/submissions',
+    const { db } = this.paramsFor('db') as { db: string };
+
+    const { content, response } = await this.requestManager.request<SubmissionRequestSummary>({
+      url: `/${db}/submission_requests`,
       options: { params: { page } },
     });
 
     return {
-      submissions: content,
+      db,
+      requests: content,
       totalPages: Number(response?.headers?.get('Total-Pages')) || 1,
     };
   }
