+++ web/app/routes/db/submissions/index.ts
@@ -4,8 +4,7 @@ import { service } from '@ember/service';
 import type { RequestManager } from '@warp-drive/core';
 import type { paths } from 'schema/openapi';
 
-type SubmissionRequestSummary =
-  paths['/{db}/submission_requests']['get']['responses']['200']['content']['application/json'];
+type SubmissionSummary = paths['/{db}/submissions']['get']['responses']['200']['content']['application/json'];
 
 export default class extends Route {
   @service declare requestManager: RequestManager;
@@ -17,13 +16,16 @@ export default class extends Route {
   };
 
   async model({ page }: { page?: number }) {
-    const { content, response } = await this.requestManager.request<SubmissionRequestSummary>({
-      url: '/st26/submission_requests',
+    const { db } = this.paramsFor('db') as { db: string };
+
+    const { content, response } = await this.requestManager.request<SubmissionSummary>({
+      url: `/${db}/submissions`,
       options: { params: { page } },
     });
 
     return {
-      requests: content,
+      db,
+      submissions: content,
       totalPages: Number(response?.headers?.get('Total-Pages')) || 1,
     };
   }
