+++ web/app/routes/request.gts
@@ -11,10 +11,12 @@ export default class extends Route {
   @service declare requestManager: RequestManager;
 
   async model({ request_id }: { request_id: string }) {
+    const { db } = this.paramsFor('db') as { db: string };
+
     const { content } = await this.requestManager.request<SubmissionRequest>({
-      url: `/st26/submission_requests/${request_id}`,
+      url: `/${db}/submission_requests/${request_id}`,
     });
 
-    return content;
+    return { db, ...content };
   }
 }
