+++ web/app/routes/update.ts
@@ -11,10 +11,12 @@ export default class extends Route {
   @service declare requestManager: RequestManager;
 
   async model({ update_id }: { update_id: string }) {
+    const { db } = this.paramsFor('db') as { db: string };
+
     const { content } = await this.requestManager.request<SubmissionUpdate>({
-      url: `/st26/submission_updates/${update_id}`,
+      url: `/${db}/submission_updates/${update_id}`,
     });
 
-    return content;
+    return { db, ...content };
   }
 }
