+++ web/app/templates/db/requests/index.gts
@@ -1,18 +1,19 @@
 import { LinkTo } from '@ember/routing';
+import { array } from '@ember/helper';
 
 import Pagination from 'repository/components/pagination';
 import StatusBadge from 'repository/components/status-badge';
 import formatDatetime from 'repository/helpers/format-datetime';
 
-import type Controller from 'repository/controllers/requests/index';
+import type Controller from 'repository/controllers/db/requests/index';
 import type { TOC } from '@ember/component/template-only';
 import type { components } from 'schema/openapi';
 
 export default <template>
-  <h1 class="display-6 mb-4">Requests</h1>
+  <h1 class="display-6 mb-4">Requests ({{@model.db}})</h1>
 
   <div class="mb-3">
-    <LinkTo @route="requests.new" class="btn btn-primary">New Submission Request</LinkTo>
+    <LinkTo @route="db.requests.new" @model={{@model.db}} class="btn btn-primary">New Submission Request</LinkTo>
   </div>
 
   <table class="table border">
@@ -28,7 +29,7 @@ export default <template>
       {{#each @model.requests as |request|}}
         <tr>
           <td>
-            <LinkTo @route="request" @model={{request.id}}>
+            <LinkTo @route="request" @models={{array @model.db request.id}}>
               Request-{{request.id}}
             </LinkTo>
           </td>
@@ -40,10 +41,16 @@ export default <template>
     </tbody>
   </table>
 
-  <Pagination @route="requests.index" @current={{@controller.page}} @total={{@model.totalPages}} />
+  <Pagination
+    @route="db.requests.index"
+    @models={{array @model.db}}
+    @current={{@controller.page}}
+    @total={{@model.totalPages}}
+  />
 </template> satisfies TOC<{
   Args: {
     model: {
+      db: string;
       requests: components['schemas']['SubmissionRequestSummary'][];
       totalPages: number;
     };
