+++ web/app/templates/db/requests/new.gts
@@ -16,7 +16,13 @@ import type { paths } from 'schema/openapi';
 type CreateRequestResponse =
   paths['/{db}/submission_requests']['post']['responses']['202']['content']['application/json'];
 
-export default class extends Component {
+interface Signature {
+  Args: {
+    model: { db: string };
+  };
+}
+
+export default class extends Component<Signature> {
   @service declare requestManager: RequestManager;
   @service declare router: RouterService;
 
@@ -33,6 +39,7 @@ export default class extends Component {
 
     if (!this.file) return;
 
+    const { db } = this.args.model;
     const upload = new DirectUpload(this.file, ENV.directUploadURL);
 
     const blob = await new Promise<Blob>((resolve, reject) => {
@@ -40,16 +47,16 @@ export default class extends Component {
     });
 
     const { content } = await this.requestManager.request<CreateRequestResponse>({
-      url: '/st26/submission_requests',
+      url: `/${db}/submission_requests`,
       method: 'POST',
       data: { submission_request: { ddbj_record: blob.signed_id } },
     });
 
-    this.router.transitionTo('request', content.id);
+    this.router.transitionTo('request', db, content.id);
   }
 
   <template>
-    <h1 class="display-6 mb-4">New Request</h1>
+    <h1 class="display-6 mb-4">New Request ({{@model.db}})</h1>
 
     <form {{on "submit" this.submit}}>
       <div class="mb-3">
