+++ web/app/templates/request.gts
@@ -3,6 +3,7 @@ import { LinkTo } from '@ember/routing';
 import { action } from '@ember/object';
 import { on } from '@ember/modifier';
 import { service } from '@ember/service';
+import { array } from '@ember/helper';
 
 import { eq } from 'ember-truth-helpers';
 
@@ -17,7 +18,7 @@ import type { components } from 'schema/openapi';
 
 interface Signature {
   Args: {
-    model: components['schemas']['SubmissionRequest'];
+    model: { db: string } & components['schemas']['SubmissionRequest'];
   };
 }
 
@@ -30,7 +31,7 @@ export default class extends Component<Signature> {
     const { model } = this.args;
 
     await this.requestManager.request({
-      url: `/st26/submission_requests/${model.id}/submission`,
+      url: `/${model.db}/submission_requests/${model.id}/submission`,
       method: 'POST',
     });
 
@@ -120,7 +121,7 @@ export default class extends Component<Signature> {
       {{#if @model.submission}}
         <h2>Submission</h2>
 
-        <LinkTo @route="submission" @model={{@model.submission.id}}>
+        <LinkTo @route="submission" @models={{array @model.db @model.submission.id}}>
           Submission-{{@model.submission.id}}
         </LinkTo>
       {{/if}}
