+++ web/tests/acceptance/submission-request-test.gts
@@ -27,17 +27,17 @@ module('Acceptance | submission request', function (hooks) {
       }),
     );
 
-    await visit('/requests');
+    await visit('/st26/requests');
 
-    assert.strictEqual(currentURL(), '/requests');
-    assert.dom('h1').hasText('Requests');
+    assert.strictEqual(currentURL(), '/st26/requests');
+    assert.dom('h1').hasText('Requests (st26)');
 
     // --- Navigate to new request page ---
 
-    await click('a[href="/web/requests/new"]');
+    await click('a[href="/web/st26/requests/new"]');
 
-    assert.strictEqual(currentURL(), '/requests/new');
-    assert.dom('h1').hasText('New Request');
+    assert.strictEqual(currentURL(), '/st26/requests/new');
+    assert.dom('h1').hasText('New Request (st26)');
 
     // --- Upload file and submit ---
 
@@ -83,7 +83,7 @@ module('Acceptance | submission request', function (hooks) {
     await triggerEvent('input[type="file"]', 'change', { files: [file] });
 
     await click('button[type="submit"]');
-    await waitUntil(() => currentURL() === '/requests/42');
+    await waitUntil(() => currentURL() === '/st26/requests/42');
 
     // --- Request detail page (ready to apply) ---
     assert.dom('h1').hasText('Request-42');
